+++ src/characters.html
@@ -15,10 +15,6 @@
 </head>
 
 <body class="background">
-    <!-- DATA -->
-
-    <script src="./data/rickandmorty/rickandmorty.js"></script>
-
 
     <header>
         <div class="main-nav">
@@ -75,7 +71,7 @@
 
         </section>
         <div class="loading-indicator">
-            <i class="fas fa-spinner Icon-size"></i> LOADING 
+            <i class="fas fa-spinner Icon-size"></i> LOADING
         </div>
     </div>
     <footer class="footer">
@@ -92,6 +88,7 @@
                 </div>
     </footer>
     <script src="main.js" type="module"></script>
+    <script src="characters.js" type="module"></script>
 </body>
 
 </html>
